+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@
 
 Changes:
 
+  * Show included projects in Runs list
   * Allow archiving labs and institutes
 
 # 1.1.0